Working Popup handler with windwo create and destroy dynamically approach.
It seems this is the only and correct way. I could not resolve the issue of the first commit.

One issue remianed: when I log in in hypothes.is bookmarklet, and if I leave the website to google for example if I return to the bookmarklet, it needs again sign in. It should remember my last sign in. This is what I will try to do in the next commit.